Reading the Config file works.
The Config::get() fn uses the new dependency `directories-next` to get the correct config file location, depending on the platform. if there is no file named after CONFIG_FILE_NAME at the Location, it creates it and writes the default config (included in the binary). If this also fails, it uses the DEFAULT_CONFIG as a fallback. After, it parses the file as JSON using `serde` and returns.