implemented the TGM3 Algorithm for drawing new pieces.
The spawn function of the TetrominoSpawner now takes a &mut self, changing the trait. The TgmBag holds the state of the current bag and handles drawing new tetrominos using the draw function. I have removed the itertools dependency. The TetrominoSpawner struct now holds a TgmBag and now uses a new() fn instead of instantiation to generate a valid state for the bag.